move reach env to separate module
